+++ src/main/java/net/minecraft/client/particle/EntityFishWakeFX.java
@@ -46,8 +46,8 @@ public class EntityFishWakeFX extends EntityFX {
 		this.prevPosX = this.posX;
 		this.prevPosY = this.posY;
 		this.prevPosZ = this.posZ;
-		this.motionY -= (double) this.particleGravity;
-		this.moveEntity(this.motionX, this.motionY, this.motionZ);
+        this.motionY -= this.particleGravity;
+        this.moveEntity(this.motionX, this.motionY, this.motionZ);
 		this.motionX *= 0.9800000190734863D;
 		this.motionY *= 0.9800000190734863D;
 		this.motionZ *= 0.9800000190734863D;
