+++ src/main/java/net/minecraft/client/particle/EntityRainFX.java
@@ -48,8 +48,8 @@ public class EntityRainFX extends EntityFX {
 		this.prevPosX = this.posX;
 		this.prevPosY = this.posY;
 		this.prevPosZ = this.posZ;
-		this.motionY -= (double) this.particleGravity;
-		this.moveEntity(this.motionX, this.motionY, this.motionZ);
+        this.motionY -= this.particleGravity;
+        this.moveEntity(this.motionX, this.motionY, this.motionZ);
 		this.motionX *= 0.9800000190734863D;
 		this.motionY *= 0.9800000190734863D;
 		this.motionZ *= 0.9800000190734863D;
@@ -74,9 +74,9 @@ public class EntityRainFX extends EntityFX {
 		if (material.isLiquid() || material.isSolid()) {
 			double d0 = 0.0D;
 			if (iblockstate.getBlock() instanceof BlockLiquid) {
-				d0 = (double) (1.0F - BlockLiquid
-						.getLiquidHeightPercent(((Integer) iblockstate.getValue(BlockLiquid.LEVEL)).intValue()));
-			} else {
+                d0 = 1.0F - BlockLiquid
+                        .getLiquidHeightPercent(iblockstate.getValue(BlockLiquid.LEVEL).intValue());
+            } else {
 				d0 = block.getBlockBoundsMaxY();
 			}
