+++ src/main/java/net/minecraft/nbt/NBTTagFloat.java
@@ -95,7 +95,7 @@ public class NBTTagFloat extends NBTBase.NBTPrimitive {
 	}
 
 	public double getDouble() {
-		return (double) this.data;
+		return this.data;
 	}
 
 	public float getFloat() {
